<feat>[chart module] include in group model
